% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Balaban_index_calc}
\alias{Balaban_index_calc}
\title{Moleculors Balaban index calculator}
\usage{
Balaban_index_calc()
}
\value{
Baladan index. Value is stored in Ouput_descp environment.
}
\description{
This function take as input the Vertex adjacency matrix, the Vertex
distance matrix,the Edge adjacency matrix and compute the
Balaban index as Edges/ 'Edges - Vertex + 1' x summatory 'total_distancei * total_distancej'^-0.5
for any adjacence vertex. The number 'Edges - Vertex + 1' is denominated with u and is
called cyclomatic number
}
\examples{
Balaban_index_calc()

}
