% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Bulk_Electronegativity_indexes_calc}
\alias{Bulk_Electronegativity_indexes_calc}
\title{Moleculors Molecular bulk indexes and Electronegativity indexes}
\usage{
Bulk_Electronegativity_indexes_calc()
}
\value{
13 different indexes regarding: bulk and electronic distribution.  Values are stored in Ouput_descp environment.
}
\description{
This function define the molecular bulk of the selected molecule as
the summatory of alpha, with alpha Z - Zv / Z * 1/ PN -1
where Z is the atomic number of the selected atom, Zv is the number of
valence electrons and PN is the periodic number of the atom.
The comparison of alpha summatory with a reference molecules
without heteroatoms allows the calculation of two other indexes
dalphaA = 'summatory alpha - summatory alpha reference' / Natoms without hydrogens
This index gives informations about the presence of heteroatoms
dalphaB = 'sumamtory alpha reference - summatory alpha' /Natoms without hydrogens
This index gives a rough count of hydrogen bond acceptor atoms and polar surface areas.
}
\details{
This function defines also the electronegativity index and other correlated indexes.
epsilon = - alpha + 0.3* Zv Electronegativity of each atoms
epsilon1 = summatory epsilon / Nof atoms including hydrogens. A measure of electronegativity atoms count
epsilon2 = summatory of epsilon excluding hydrogens / Nof atoms excluding hydrogens
epsilon3 = summatory of epsilon reference / Nof reference atoms
epsilon4 = summatory of epsilon of the saturated carbon skeleton /Nof atoms saturated carbon skeleton
epsilon5 = summatory of epsilon excluding hydrogens + summatory of epsilon of hydrogens attached to heteroatoms / Nof atoms excluding hydrogens + Nof hydrogens attached to heteroatoms

DepsilonA =epsilon1 - epsilon3 measure of contribution of unsaturation and electronegative atom count
DepsilonB = epsilon1 - epsilon4 measure of contribuion of unsaturation
DepsilonC = epsilon3 - epsilon4 a measure of contribuion of electronegativity
DepsilonD = epsilon2 - epsilon5 a measure of contribution of hydrogen-bond donor atoms
}
\examples{
Bulk_Electronegativity_indexes_calc()

}
