% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Bulk_index_calc}
\alias{Bulk_index_calc}
\title{Moleculors Molecular bulk index}
\usage{
Bulk_index_calc()
}
\value{
Molecular bulk for the selected molecule. Value is stored in Ouput_descp environment.
}
\description{
This function define the molecular bulk of the selected molecule as
the summatory of Z - Zv / Z * 1/ PN -1
where Z is the atomic number of the selected atom, Zv is the number of
valence electrons and PN is the periodic number of the atom.
The only required matrix is the input_H_suppressed as Hydrogen volume is taken
as reference and thus its value is set to zero.
}
\examples{
Bulk_index_calc()

}
