% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Distance_path_Vdistance}
\alias{Distance_path_Vdistance}
\title{Moleculors vertex distance path matrix}
\usage{
Distance_path_Vdistance()
}
\value{
Vertex distance path matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function compute the vertex version of the distance-path-matrix
It take has input the Vdistance_matrix and for each
element of the matrix if i != j it returns elemij*'elemij + 1'/2 value of
the distance matrix
}
\examples{
Distance_path_Vdistance()

}
