% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{E_Harary_matrix}
\alias{E_Harary_matrix}
\title{Moleculors edge harary matrix}
\usage{
E_Harary_matrix()
}
\value{
Edge Harary matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function compute the edge version of the so called
Harary matrix. It take has input the Edistance_matrix and for each
element of the matrix if i != j it returns the reciprocal value of
the distance matrix
}
\examples{
E_Harary_matrix()

}
