% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{E_state}
\alias{E_state}
\title{Moleculors Electrotopological state index}
\usage{
E_state()
}
\value{
E-state index for the selected molecule. Value is stored in Ouput_descp environment.
}
\description{
This function requires access to the vertex adjacency matrices, full and suppressed, to the
raw inputs, full and suppressed, to the vertex laplacian matrix and to the vertex distance matrix
in order to compute the E-state index for the whole molecule. This is computed by calculating first
the intrinsic factor for each atom as 2/principal quanto number ^2 x valence degree + 1 / degree of the atom
and correcting this factor by a Di equal to the summatory of intrinsic state of atom i - intrinsic state of atom j
over the topological distanze r + 1^2. The obtained E-state index I + Di can be used for the single atoms of can be summed
to get the E-state of substituent or, as was done in this function, to compute the whole molecule E-state
}
\examples{
E_state()

}
