% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Eadj_matrix}
\alias{Eadj_matrix}
\title{Moleculors edge adjacency matrix}
\usage{
Eadj_matrix(Cart_Input_Hsupp)
}
\arguments{
\item{Cart_Input_Hsupp}{Cartesian coordinates of the molecule without the hydrogen atoms}
}
\value{
Edge adjacency matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function require the H suppressed cartesian matrix
and the adjagency matrix for the vertexes in order to compute
the edge cartesian matrix and the adjacency edges matrices.
}
\examples{
Eadj_matrix(input_H_suppressed)

}
