% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Eadj_matrix_full}
\alias{Eadj_matrix_full}
\title{Moleculors edge adjacency matrix full}
\usage{
Eadj_matrix_full(Cart_Input)
}
\arguments{
\item{Cart_Input}{Cartesian coordinates of the molecule without the hydrogen atoms}
}
\value{
Edge adjacency matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function require the cartesian matrix
and the adjacency matrix for the vertexes in order to compute
the edge cartesian matrix and the adjacency edges matrices.
}
\examples{
Eadj_matrix_full(input)

}
