% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Ec_index_calc}
\alias{Ec_index_calc}
\title{Moleculors eccentric connectivity index}
\usage{
Ec_index_calc()
}
\value{
eccentric connectivity index for the input molecule. Value is stored in Ouput_descp environment.
}
\description{
This function define the eccentric connectivity index of the input molecule as:
Ec_index = sum"csii*sigmai" where csii is the max topological distance from the atom
i and sigmai is is vertex degree
}
\examples{
Ec_index_calc()

}
