% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Edistance_matrix}
\alias{Edistance_matrix}
\title{Moleculors edge distance matrix}
\usage{
Edistance_matrix()
}
\value{
Edge distance matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function takes as input the Eadjancency matrix and compute, using
the same algorithm used for the vertexes, the distance matrix for the edges.
This is done by computing a connection vector that states with which edge each edge is connected
than it loop through each connection to find the wanted connection 'e.g. elemij It loop through
the connection of 1 until it find the element j, increasing the distance at each loop'
}
\examples{
Edistance_matrix()

}
