% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{eigenvalues_descp_cacl}
\alias{eigenvalues_descp_cacl}
\title{Moleculors adj matrix eigenvalues descriptors}
\usage{
eigenvalues_descp_cacl()
}
\value{
eigenvalues descriptors for the input molecule. Value is stored in Ouput_descp environment.
}
\description{
This function define several descriptors from the eigenvalue of the adjmatrix:
Epi represents the energy of the energy of the pi electrons of the molecule. It is define as the
summatory of the eigenvalue i times the occupation number of the MO i, this can be 0,1 or 2. Since usually
the orbital contains 2 electrons when the eigenvalue is positive and 0 when negative. We can redefine the Epi
value as the summatory of the absolute value of the eigenvalue. Since the number of eigenvalues is equal to the number of
atoms in the H depleted matrxi, if the number is ODD we will have 1 occupation number, this is automatically taken into account
when adding the absolute values. e.g. Epi = 2 x +y1 + 2 x +y2 + 1 x +y3 + 0 x -y4 0 x -y5
LEDM "leading value of distance matrix" represent the max value of the eigenvalue of the Vdistance matrix.
This descriptor can be a discriminant of molecules size in molecules series.
Sparse CRI "characteristic root index" represent the sum of the positive eigenvalues of the sparse csi
matrix. Such value is sensitive to the presence of heteroatoms in the molecule.
Extadj_Vdegree_index and Extadj_Vdegree_max_index represent respectively the sum of the eigenvalues and the
maximum value of the eigenvalues of the extended Vadj matrix for vertex degree. This descriptors are
usually well correlated with physico-chemical properties and biological activities of organic compounds.
Extadj_Veln_index and Extadj_Veln_max_index represent respectively the sum of the eigenvalues and the
maximum value of the eigenvalues of the extended Vadj matrix for vertex electronegativity. This descriptors are
usually well correlated with physico-chemical properties and biological activities of organic compounds and affected by
the presence of heteroatoms.
}
\examples{
eigenvalues_descp_cacl()

}
