% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Extended_Vadj_matrix_Vdegrees}
\alias{Extended_Vadj_matrix_Vdegrees}
\title{Moleculors extended adjacency matrix from vertex degrees}
\usage{
Extended_Vadj_matrix_Vdegrees()
}
\value{
extended adjacency matrix from vertex degree for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function return the extended adjacency matrix taking into account the
effect of vertexes degrees.
it take as input the Vadj_matrix, the laplacian_Vertex matrix
and for each element return a'ij = ai * 'di/dj + dj/di'/2 where di is the vertex degree
of the element i. By definition a'ij with i=j will be 0
}
\examples{
Extended_Vadj_matrix_Vdegrees()

}
