% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Extended_Vadj_matrix_Veln}
\alias{Extended_Vadj_matrix_Veln}
\title{Moleculors extended adjacency matrix from vertex electronegativity}
\usage{
Extended_Vadj_matrix_Veln()
}
\value{
extended adjacency matrix from vertex electronegativity for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function return the extended adjacency matrix taking into account the
effect of vertexes electronegativity.
it take as input the Vadj_matrix, the laplacian_Vertex matrix
and for each element return a'ij = ai * 'di''/dj'' + dj''/di''/2 where di'' is the vertex degree
of the element i multiplied by its electronegativity elni. By definition a'ij with i=j will be elni
}
\examples{
Extended_Vadj_matrix_Veln()

}
