% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{First_order_kappa_calc}
\alias{First_order_kappa_calc}
\title{Moleculors Kappa shape index first order}
\usage{
First_order_kappa_calc()
}
\value{
First order kappa shape index. Value is stored in Ouput_descp environment.
}
\description{
This function takes the vertex adjacency matrix and the edge adjacency matrix and
return the first order Kappa shape index. This parameters gives information about the
shape of the molecule. As K is calculated as Pmin * Pmax /P where Pmin = V -1
Pmax = V * 'V-1' and P is the number of paths of the molecules aka edges.
This parameter gives information about the numbers of cycles inside the molecule. The lower
K compared to V the more cycle are present.
}
\examples{
First_order_kappa_calc()

}
