% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algebraic_utilities.R
\name{gauss_jordan_inverse}
\alias{gauss_jordan_inverse}
\title{Moleculors inverse matrix with gauss jordan algorithm}
\usage{
gauss_jordan_inverse(A)
}
\value{
Inverse matrix of the input matrix A
}
\description{
Compute the inverse matrix of the input A using the gauss jordan algorithm
Given A*X-1 = I we solve with gauss algorithm the augmented matrix 'AI'
by forward elimination and backward we can manage to obtain 'IX-1'
}
\examples{
A = matrix(1,2,3,5)
invMat(A)

}
