% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Hydro_factor_calc}
\alias{Hydro_factor_calc}
\title{Moleculors hydrophilicity factor}
\usage{
Hydro_factor_calc()
}
\value{
Hydrophilicity factor for the input molecule. Value is stored in Ouput_descp environment.
}
\description{
This function define the hydrophilicity factor of the input molecule as:
Hy = ""1+Nhy" *log2 "1+Nhy" + Nc "1/A log2 1/A" + sqrt"Nhy/A^2""/log"1+A"
where Nhy is the number of hydrogen attached to oxygen sulfur or nitrogen groups,
Nc is the number of carbons and A is the number of non hydrogen atoms
}
\examples{
Hydro_factor_calc()

}
