% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algebraic_utilities.R
\name{invMat_old_not_supported}
\alias{invMat_old_not_supported}
\title{Moleculors inverse matrix}
\usage{
invMat_old_not_supported(A)
}
\value{
Inverse matrix of the input matrix A
}
\description{
Compute the inverse matrix of the input A using the determinant and the adj matrix
this function is not computing efficient due to the computation of the determinant
is supposed to be used for small matrices
}
\examples{
A = matrix(1,2,3,5)
invMat_old_not_supported(A)

}
