% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Laplacian_Vdistance}
\alias{Laplacian_Vdistance}
\title{Moleculors vertex Laplacian matrix}
\usage{
Laplacian_Vdistance()
}
\value{
Laplacian vertex matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function return the laplacian matrix for the vertex distance matrix
it take as input the Vdistance_matrix and for each element return
-1 if elemij = 1, sum'elemi, == 1' if i = j 'which is the angle of the vertex'
and 0 for every other element
}
\examples{
Laplacian_Vdistance()

}
