% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{MED_index_calc}
\alias{MED_index_calc}
\title{Moleculors mean electronic distribution index}
\usage{
MED_index_calc()
}
\value{
mean electronic distribution index. Value is stored in Ouput_descp environment.
}
\description{
This function define the mean electronic distribution as follow:
mED = 'number of coniugated double bonds/ number of atoms in the H suppresed matrix' +
'number of electron donators groups'/number of the double bonds * max distance from the double bonds  -
'number of electron actractor groups'/number of the double bonds * max distance from the double bonds
a donator group is: C-C/OH/OR/NH2/NHR/NR2/SH/SR while an actractor is:
NO2/CN/F/Cl/Br/I/C---C
}
\examples{
Unsaturation_index_calc()

}
