% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moleculors_V01.R
\docType{data}
\name{Mol_mat}
\alias{Mol_mat}
\title{Matrices environment}
\format{An object of class \code{environment} of length 0.}
\usage{
Mol_mat
}
\description{
Matrices environment
}
\keyword{datasets}
