% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_Input_v01.R
\name{molecular_input}
\alias{molecular_input}
\title{Moleculors input file loader.}
\usage{
molecular_input()
}
\description{
This function take a csv file containing 4 colums:
atom symbol, X,Y,Z. Any difference will be detected as
an error or warning depending on the situation. NOTE: the input
coordinates are supposed to be the optimized coordinates of the
target molecule! If this requirement is not fullfilled any descriptor
will not be related to the topological property at all.
}
\examples{
molecular_input()


}
