% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Moleculors$Complementary_Edistance}
\alias{Moleculors$Complementary_Edistance}
\title{Moleculors edge complementary matrix}
\usage{
Moleculors$Complementary_Edistance()
}
\value{
Edge complementary matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function compute the edge version of the complementary
edge matrix . It take has input the Edistance_matrix and for each
element of the matrix if i != j it returns the  value of dmin + dmax - elemij
were dmin is the minimal lenght in the molecular graph (1 for simple molecules)
dmax is the maximum lenght in the system and elemij is the value of the element i j
of the edge distance matrix
}
\examples{
Moleculors$Complementary_Edistance()

}
