% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Moleculors$Complementary_Vdistance}
\alias{Moleculors$Complementary_Vdistance}
\title{Moleculors vertex complementary matrix}
\usage{
Moleculors$Complementary_Vdistance()
}
\value{
Vertex reciprocal matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function compute the vertex version of the complementary
vertex matrix . It take has input the Vdistance_matrix and for each
element of the matrix if i != j it returns the  value of dmin + dmax - elemij
were dmin is the minimal lenght in the molecular graph (1 for simple molecules)
dmax is the maximum lenght in the system and elemij is the value of the element i j
of the vertex distance matrix
}
\examples{
Moleculors$Complementary_Vdistance()

}
