% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptor_launcher_V01.R
\name{Moleculors$descriptor_launcher}
\alias{Moleculors$descriptor_launcher}
\title{Moleculors descriptor calculator launcher}
\usage{
Moleculors$descriptor_launcher()
}
\value{
Returns the molecular descriptors '0D, 1D, 2D' for the selected molecule. Results are stored in Output_descp environment.
}
\description{
This function runs the Molecular descriptors functions and handle each of
them automatically. The user is suggested to start this function rather than each one
of the manually if all the descriptors are required.
}
\examples{
Moleculors$descriptor_launcher()

}
