% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Moleculors$Distance_path_Edistance}
\alias{Moleculors$Distance_path_Edistance}
\title{Moleculors edge distance path matrix}
\usage{
Moleculors$Distance_path_Edistance()
}
\value{
Edge distance path matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function compute the edge version of the distance-path-matrix
It take has input the Edistance_matrix and for each
element of the matrix if i != j it returns elemij*'elemij + 1'/2 value of
the distance matrix
}
\examples{
Moleculors$Distance_path_Edistance()

}
