% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Moleculors$ECdistance_matrix}
\alias{Moleculors$ECdistance_matrix}
\title{Moleculors edge complement distance matrix}
\usage{
Moleculors$ECdistance_matrix()
}
\value{
Edge complement distance matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function takes as input the Edistance matrix and compute for each element
the complement as nrowEdistance_matrix - Edistance_matrixi,j
}
\examples{
Moleculors$ECdistance_matrix()

}
