% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Moleculors$graphical_matrix}
\alias{Moleculors$graphical_matrix}
\title{Moleculors graphical matrices calculator}
\usage{
Moleculors$graphical_matrix()
}
\value{
Graphical matrices for edges and verteces stored inside the Mol_mat environment.
}
\description{
This function take the cartesian coordinates of a molecule and return the graphical matrices
of its structure. To do this first of all remove all the hydrogens leaving only the carbons
and heteroatoms. Then it compute the relative distance for each atom from each other and
round it up. In the end we will have a matrix n*n where n is the number of atoms
and each cell contain an integer pointing to the number of bonds intercurring
between the diagonal element and the others. The adjacency matrix is then computed by checking if
the distance is 1 or not.
For the edge matrix, the edge coordinate is computed from the input coordinates as the mid point
between two adjacent atoms and then the distance and adjacency is computed using the previuos
algorithm
}
\examples{
Moleculors$graphical_matrix()

}
