% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Moleculors$Laplacian_Edistance}
\alias{Moleculors$Laplacian_Edistance}
\title{Moleculors edge Laplacian matrix}
\usage{
Moleculors$Laplacian_Edistance()
}
\value{
Laplacian edge matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function return the laplacian matrix for the edge distance matrix
it take as input the Edistance_matrix and for each element return
-1 if elemij = 1, sum'elemi, == 1' if i = j 'which is the angle of the edge'
and 0 for every other element
}
\examples{
Moleculors$Laplacian_Edistance()

}
