% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroD_descriptor_V01.R
\name{Moleculors$molecular_weight}
\alias{Moleculors$molecular_weight}
\title{Moleculors Molecular weight descriptor}
\usage{
Moleculors$molecular_weight()
}
\value{
Molecular weight of the selected molecule. Values are store inside Output_descp environment.
}
\description{
this function compute the molecular weight of the input molecule
by checking the atom symbol in the input matrix/df and crossing it
with an internal library containing atoms weight.
}
\examples{
Moleculors$molecular_weight()

}
