% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Moleculors$Platt_number_calc}
\alias{Moleculors$Platt_number_calc}
\title{Moleculors Platt number calculator}
\usage{
Moleculors$Platt_number_calc()
}
\value{
Platt number. Value is stored in Ouput_descp environment.
}
\description{
This function take as input the edge adjacency matrix and return the Platt Number
this number is calculated as the summatory of the summatory of each elementij in each row
of the adjacency matrix
}
\examples{
Moleculors$Platt_number_calc()

}
