% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Moleculors$Vdistance_matrix}
\alias{Moleculors$Vdistance_matrix}
\title{Moleculors vertex distance matrix}
\usage{
Moleculors$Vdistance_matrix()
}
\value{
Vertex distance matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function take the adjacency matrix as input and generate a connection list where the index
is related to the atom and the value in the index list are the atoms to which that atom is connected
For each element with d != 1 is then calculated the distance by taking the value in the connection list
copying those into the connection vector and if the interested atom j is not in the connection vector
a new vector is created with all the connection of the atom to which the first atoms was connected
this is looped increasing distance at each failed loop to detect the interested element.
}
\examples{
Moleculors$Vdistance_matrix()

}
