% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Moleculors$Zagreb_index_calc}
\alias{Moleculors$Zagreb_index_calc}
\title{Moleculors Zagreb index calculator}
\usage{
Moleculors$Zagreb_index_calc()
}
\value{
Zagreb index. Value is stored in Ouput_descp environment.
}
\description{
This function take as input the Laplacian Vertex matrix and return
the Zagreb index for the selected molecular graph as the summatory
of the squared degree for each atom of the matrix
}
\examples{
Moleculors$Zagreb_index_calc()

}
