% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moleculors_V01.R
\docType{data}
\name{Moleculors}
\alias{Moleculors}
\title{include Molecular_Input_v01.R Molecular_graph_V01.R GUI_V01.R zeroD_descriptor_V01.R twoD_descriptor_V01.R descriptor_launcher_V01.R
Moleculors environment}
\format{An object of class \code{environment} of length 11.}
\usage{
Moleculors
}
\description{
include Molecular_Input_v01.R Molecular_graph_V01.R GUI_V01.R zeroD_descriptor_V01.R twoD_descriptor_V01.R descriptor_launcher_V01.R
Moleculors environment
}
\keyword{datasets}
