% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_V01.R
\name{Moleculors_GUI}
\alias{Moleculors_GUI}
\title{Moleculors GUI}
\usage{
Moleculors_GUI()
}
\value{
Graphical user interface for Moleculors functions
}
\description{
Basic GUI for moleculors functions User is not supposed to use this functions. GUI should be loaded using Moleculors_init
}
\examples{
Moleculors_GUI()

}
