% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moleculors_V01.R
\name{Moleculors_init}
\alias{Moleculors_init}
\title{include Molecular_Input_v01.R Molecular_graph_V01.R GUI_V01.R zeroD_descriptor_V01.R twoD_descriptor_V01.R descriptor_launcher_V01.R
Moleculors initialization and environments set.}
\usage{
Moleculors_init(GUI = TRUE)
}
\arguments{
\item{GUI}{Graphical user interface option. If true launch the GUI.}
}
\value{
if GUI = TRUE Moleculors environment is initializated with a GUI, else just set the environment
}
\description{
Every functions will be part and only part
of this environment thus preventing global variables overlay. Graphic
matrices will belong to the Mol_mat envinroment and molecular descriptor to
Ouput_descp
}
\examples{
Moleculors_init()

}
