% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lin_reg.R
\name{Moleculors_lin_reg}
\alias{Moleculors_lin_reg}
\title{Moleculors linear regression}
\usage{
Moleculors_lin_reg(descp, output, row_wise = T, split_ratio = 0.7)
}
\value{
Linear regression model of the descriptors and output.
}
\description{
Compute a linear model of the descriptors and selected output. Data are automatically splitted
in train/validation with default 0.7 ratio.
}
\examples{
Moleculors_lin_reg()

}
