% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_Input_v01.R
\name{moleculors_multiple_descriptor}
\alias{moleculors_multiple_descriptor}
\title{Moleculors multiple graph and descriptor calculator.}
\usage{
moleculors_multiple_descriptor()
}
\description{
This function loop through the multiple input selected and calculate
the graphical matrices and descriptor for each molecule. Results are stored
in the descriptor matrix where each column is a molecule and each row are the
orded descriptors
}
\examples{
moleculors_multiple_descriptor()


}
