% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroD_descriptor_V01.R
\name{N_atoms}
\alias{N_atoms}
\title{Moleculors number of atoms descriptor}
\usage{
N_atoms()
}
\value{
Number of atoms in the selected molecule. Values are store inside Output_descp environment.
}
\description{
This function calculate the number of atoms of the
molecular input by simply calculating the number of rows of the input matrix
}
\examples{
N_atoms()

}
