% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algebraic_utilities.R
\name{OneMat}
\alias{OneMat}
\title{Moleculors Ones Matrix}
\usage{
OneMat(i, j = i)
}
\value{
Generates a ixj matrix filled with ones.
}
\description{
Generates a ixj matrix filled with ones
}
\examples{
OneMat(2)

}
