% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moleculors_V01.R
\docType{data}
\name{Output_descp}
\alias{Output_descp}
\title{Descriptors environment}
\format{An object of class \code{environment} of length 0.}
\usage{
Output_descp
}
\description{
Descriptors environment
}
\keyword{datasets}
