% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_Input_v01.R
\name{prediction_parameter}
\alias{prediction_parameter}
\title{prediction matrix loader.}
\usage{
prediction_parameter()
}
\description{
This function take as input the prediction vector to be used
for the training of the NN for the qsar study and store the information in
the prediction matrix
}
\examples{
prediction_parameter()


}
