% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Q_polarity_calc}
\alias{Q_polarity_calc}
\title{Moleculors Q polarity index}
\usage{
Q_polarity_calc()
}
\value{
Q polarity index for the input molecule. Value is stored in Ouput_descp environment.
}
\description{
This function define the Q polarity index of the input molecule as:
Q = ' A^2 sumIrefskeleton/sumI^2 '
where A is the numbero of atoms in the H suppresed input, Irefskeleton is the
intrinsic state of each atom in the reference alcane chain and I is the intrisic state
in the actual molecule. Q lies between to boundary value, one of total polarity approximated
by the square of the number of atoms A  and one of zero polarity given by the sp3 skeleton
}
\examples{
Q_polarity_calc()

}
