% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neural_network.R
\name{QSAR_NN_model}
\alias{QSAR_NN_model}
\title{Moleculors Neural network.}
\usage{
QSAR_NN_model(descriptor, predictor, train_split = 0.75, hl = 4, lr = 0.1)
}
\arguments{
\item{train_split}{train test split ratio for the input data}

\item{hl}{number of nodes in the hidden layer}
}
\description{
This function takes as input the descriptor matrix, the predicted values
and return the performance in prediction of a neural network with one hidden layer
the train_split argument will determine how data are divided for train/test/validation.
}
\examples{
QSAR_NN_model()


}
