% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Randic_index_calc}
\alias{Randic_index_calc}
\title{Moleculors Randic connectivity index calculator}
\usage{
Randic_index_calc()
}
\value{
Randic connectivity index. Value is stored in Ouput_descp environment.
}
\description{
This function takes the Vertex adjacency matrix and the laplacian matrix
and return the randic connectivity index calculated as the summatory of the product
of the element ij of the adjacency matrix and the inverse of square root
of the product of the degree of the vertexes i j. NOTE this index can be considered
as the Kier and Hall index of the first order.
}
\examples{
Randix_index_calc()

}
