% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Randic_valence_index_calc}
\alias{Randic_valence_index_calc}
\title{Moleculors Randic valence connectivity index calculator}
\usage{
Randic_valence_index_calc()
}
\value{
Randic valence connectivity index. Value is stored in Ouput_descp environment.
}
\description{
This function takes the Vertex adjacency matrix, the full adjacency matrix, the raw input and the input with H suppressed
and return the randic valence connectivity index calculated as the summatory of the product
of the valence corrected degree of each atom. This element calculated as Valence electrons - number of hydrogen over
total electrons of the atom - valence electrons - 1. In this way this element take into account the effect of heteroatoms
and multiple bonds.
}
\examples{
Randix_valence_index_calc()

}
