% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Rec_Complementary_Vdistance}
\alias{Rec_Complementary_Vdistance}
\title{Moleculors vertex reciprocal complementary matrix}
\usage{
Rec_Complementary_Vdistance()
}
\value{
Vertex reciprocal complementary matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function compute the vertex version of the reciprocal
complementary vertex matrix . It take has input the Complementary_Vdistance_matrix and for each
element of the matrix if i != j it returns the reciprocal value of
the distance matrix
}
\examples{
Rec_Complementary_Vdistance()

}
