% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Sparse_csi_V_matrix}
\alias{Sparse_csi_V_matrix}
\title{Moleculors sparse csi vertex matrix}
\usage{
Sparse_csi_V_matrix()
}
\value{
sparse Csi vertex matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function return the sparse csi vertex matrix for the selected molecule
it take as input the Vadj_matrix, the laplacian_Vertex matrix and the Vdistance matrix
 and for each element return wij = "di x dj"^-1/2 where di is the vertex degree
 of the element i. By definition wij with i=j will be 0
}
\examples{
Sparse_csi_V_matrix()

}
