% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Unsaturation_index_calc}
\alias{Unsaturation_index_calc}
\title{Moleculors Unsaturation Index}
\usage{
Unsaturation_index_calc()
}
\value{
Unsaturation index of the input molecule. Value is stored in Ouput_descp environment.
}
\description{
This function define the unsaturation index as follow:
UI = log2 "1 + SBO + nB" where nD is the number of bonds
and SBO is the sum of bonds order in the molecule. The H suppressed matrix is used for
the computation.
}
\examples{
Unsaturation_index_calc()

}
