% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Vadj_highpower_matrix}
\alias{Vadj_highpower_matrix}
\title{Moleculors vertex highpower matrix}
\usage{
Vadj_highpower_matrix(n)
}
\arguments{
\item{n}{integer for matrix power}
}
\value{
nth power vertex matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function takes as input the adjacency matrix and an integer n and
return the n power of the adjacency matrix.
}
\examples{
Vadj_highpower_matrix()

}
