% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Vadj_kneigh_matrix}
\alias{Vadj_kneigh_matrix}
\title{Moleculors vertex kneighbour matrix}
\usage{
Vadj_kneigh_matrix(n)
}
\arguments{
\item{n}{integer for kn neighbour}
}
\value{
K neighbour vertex matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function take as input the Vdistance matrix and an interger < then
nrow'vdistance' and return the adjmatrix for the kn neighbour
}
\examples{
Vadj_kneigh_matrix()

}
