% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{VCdistance_matrix}
\alias{VCdistance_matrix}
\title{Moleculors vertex distance matrix}
\usage{
VCdistance_matrix()
}
\value{
Vertex complement distance matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function requires the Vdistance_matrix in order to compute the complement
 of the distance matrix. Each element is set has nrowVdistance - Vdistanceij
}
\examples{
VCdistance_matrix()

}
