% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_graph_V01.R
\name{Vdistance_distance_matrix}
\alias{Vdistance_distance_matrix}
\title{Moleculors Vdistance/distance matrix}
\usage{
Vdistance_distance_matrix()
}
\value{
distance/distance matrix for the loaded molecule. Matrix is stored in Mol_mat environment.
}
\description{
This function return the Vdistance/distance matrix for the selected molecule
it takes as input the Vdistance matrix, compute the geometrical distance matrix for each element ij
and for each element return ddij = geomij/distanceij
}
\examples{
Vdistance_distance_matrix()

}
