% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algebraic_utilities.R
\name{ZeroMat}
\alias{ZeroMat}
\title{Moleculors Zeros Matrix}
\usage{
ZeroMat(i, j = i)
}
\value{
Generates a ixj matrix filled with zero.
}
\description{
Generates a ixj matrix filled with zeros
}
\examples{
ZeroMat(2)

}
