% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_descriptor_V01.R
\name{Wiener_index_calc}
\alias{Wiener_index_calc}
\title{Moleculors Weiner index calculator}
\usage{
Wiener_index_calc()
}
\value{
Weiner index. Value is stored in Ouput_descp environment.
}
\description{
This function take as input the vertex distance graphical matrix and return the Wiener Index
this index is calculated as half of the summatory of the summatory of the distances in each
row of the graphical matrix
}
\examples{
Wiener_index_calc()

}
